Refactor import paths and modify user connect type
